requirements.txt
python-telegram-bot==21.6
aiohttp
websockets
pandas
numpy
pytz
